Fixed required changes
Fixed the required changes that I could do. Not all of them, waiting for response.
