--- templates/ad/reply.html.twig
@@ -1,33 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Reply
-{% endblock %}
-
-{% block body %}
-	<div class="ui main text container">
-		<h3 class="ui dividing header">Reply to a comment</h3>
-		{{ form_start(form, { 'attr': {'class': 'ui large form'},'action': path('submitreply', {'id': id, 'parent': parent}), 'method': 'POST'}) }}
-		{% if error.current %}
-			<div class="ui negative message">
-				<i class="close icon"></i>
-				<div class="header" style="text-align: left">
-					There was a problem
-				</div>
-				{% for err in error %}
-					<p style="text-align: left">{{ err.messageTemplate }}</p>
-				{% endfor %}
-			</div>
-		{% endif %}
-		<div class="ui stacked segment">
-			<div class="field">
-				<div style="height: 5em" class="ui input reply form">
-					{{ form_widget(form.text) }}
-				</div>
-			</div>
-			<button class="ui primary submit labeled icon button" id="comment">
-				<i class="icon edit"></i>Reply</button>
-		</div>
-		<div class="ui error message"></div>
-		{{ form_end(form) }}
-	</div>
-{% endblock %}
